requirements.txt
requests==2.32.4
beautifulsoup4==4.13.4
lxml==6.0.0
openai==1.95.1
python-dotenv==1.1.1